news-hub/crawler/requirements.txt
feedparser
beautifulsoup4
html5lib
kafka-python